backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
google-generativeai==0.8.0
duckdb==1.1.0
polars==1.8.0
python-dateutil==2.9.0
langdetect==1.0.9
hashlib-additional==1.0.1
datasketch==1.6.1
scikit-learn>=1.6
sentence-transformers==3.1.0
bertopic==0.16.4
umap-learn==0.5.6
hdbscan==0.8.42
networkx==3.3
python-igraph==0.11.6
leidenalg==0.10.2
pyvis==0.3.2
ruptures==1.1.9
scipy==1.14.1
numpy==1.26.4
chromadb==0.5.7
redis==5.1.0
datamapplot>=0.4.2

# LangGraph chatbot (OpenRouter)
langgraph>=0.2.0
langchain-core>=0.2.0
openai>=1.0.0
